# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "This project is the rest API for a blog.",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:4000",
  "basePath": "/api",
  "tags": [],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/audiovisuals/{id}": {
      "get": {
        "tags": [
          "Audiovisual"
        ],
        "description": "Endpoint to get one Audiovisual",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Audiovisual"
        ],
        "description": "Endpoint to delete one Audiovisual",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/audiovisuals": {
      "get": {
        "tags": [
          "Audiovisual"
        ],
        "description": "Endpoint to get all Audiovisual",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Audiovisual"
        ],
        "description": "Endpoint to create a Audiovisual",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Audiovisual"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Audiovisual"
        ],
        "description": "Endpoint to update a Audiovisual",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Audiovisual"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/disertacions/{id}": {
      "get": {
        "tags": [
          "Disertacions"
        ],
        "description": "Endpoint to get one Disertacions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Disertacions"
        ],
        "description": "Endpoint to delete one disertacions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/disertacions": {
      "get": {
        "tags": [
          "Disertacions"
        ],
        "description": "Endpoint to get all Disertacions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Disertacions"
        ],
        "description": "Endpoint to create a Disertacions",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Disertacion"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Disertacions"
        ],
        "description": "Endpoint to update a Disertacions",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Disertacion"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/glosarios/{id}": {
      "get": {
        "tags": [
          "Glosarios"
        ],
        "description": "Endpoint to get one glosario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Glosarios"
        ],
        "description": "Endpoint to delete one glosarios",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/glosarios": {
      "get": {
        "tags": [
          "Glosarios"
        ],
        "description": "Endpoint to get all glosarios",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Glosarios"
        ],
        "description": "Endpoint to create a glosarios",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Glosario"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Glosarios"
        ],
        "description": "Endpoint to update a glosarios",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Glosario"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/resenias/{id}": {
      "get": {
        "tags": [
          "Resenias"
        ],
        "description": "Endpoint to get one Resenias",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Resenias"
        ],
        "description": "Endpoint to delete one resenias",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/resenias": {
      "get": {
        "tags": [
          "Resenias"
        ],
        "description": "Endpoint to get all Resenias",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Resenias"
        ],
        "description": "Endpoint to create a Resenias",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Resenia"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Resenias"
        ],
        "description": "Endpoint to update a Resenias",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Resenia"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Audiovisual": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "title": {
          "type": "string",
          "example": ""
        },
        "description": {
          "type": "string",
          "example": ""
        },
        "url": {
          "type": "string",
          "example": ""
        }
      }
    },
    "Glosario": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "word": {
          "type": "string",
          "example": ""
        },
        "definition": {
          "type": "string",
          "example": ""
        }
      }
    }
  }
}